remove thread wakeup race condition
When a user wished to remove threads, there existed a race condition where a thread had just marked itself as asleep (set job_id to -1), and the waking function just sent a signal to wake it up. The thread would enter sem_wait with nobody to notify it that it should be killed. This change fixes that and simplifies the code base. Instead of using signals to wake sleeping threads, we use the existing job queue to push a 'do nothing' job. This will wake a sleeping thread up and cause it to handle our flags.