Restructure project files to separate public API
This commit attempts to restructure the project to create a clear
separation between the public and internal API. This is important if we
want to support any real users and create a reliable piece of software.

Firstly, the main header file, fiber.h, was split into two. The original
fiber.h included some structs (like the main fiber_pool struct) that are
destined to change in the future. These were moved to
src/fiber_internal.h. Also, Items from job_queue.h were moved to fiber.h
to consolidate the public API.

I also moved the fifo job queue header file to the project's root
directory to make it part of the public API. It declares the job queue
operations and a helper macro. This needs to be included by users who
choose to use this queue so it cannot be internal.

There were two changes to the public API's interface: (Fiber is not in
version 1 yet so I'm still changing public things as needed)
1. fiber_pool_init returns an error code and fiber_pool pointer: This
   was done because fiber_pool became an opaque struct to hide
   implementation details. We could no longer rely on the user to
   allocate the memory for the pool and pass the pointer to us.
2. fiber_queue_operations.init returns an error code and queue pointer:
   This wasn't strictly necessary but I didn't like the idea of using
   double indirection to assign the job queue.

There were also some minor changes:
- Moved all compile time configuration to fiber.h: Before this,
  configuration was scattered throughout the codebase. This allows the
  user to configure the software in one place.
- Remove FIBER_NO_DEFAULT_QUEUE option: This was messy and sucked. If
  the user doesn't want to compile the fifo queue into the project, they
  should remove it from the compilation.
- Updated the existing tests to accomidate new API.