Split fiber.c into modules, decouple from pthread and gcc atomics
This is a very large commit with a lot of changes. The goal of this
commit was simply to clean up the codebase. There is only one change
that alters how the code logically executes.

I'll briefly describe all the minor changes:
- Switch from C11 to C89.
- Add a lot of warning flags to compilation command.
- Created a generic interface for threading utilities. These include
  mutexes, semaphores, and thread stuff. Added an implementation using
  pthreads.
- Created a generic interface for needed atomic operations. Added an
  implementation using GCC/Clang built in atomic functions.
- Cleanup up the assert statements by removing the error message param.
- Pushed a cleanup routine onto the thread's cleanup stack. This
  ensures resources are freed even if the thread is canceled. Prior
  versions of Fiber leaked memory if the thread was canceled.

The major change of this commit is splitting fiber.c into different
modules. fiber.c grew too large and had too many static functions. There
were two files created from this:
- worker.c: Worker thread functions.
- thread_list.c: Linked list helper functions for fiber_thread structs.