Fix race condition in fifo_push and fifo_pop
In the queue impl functions fiber_queue_fifo_push and fiber_queue_fifo_pop
the values of head and tail were not being guarded by any locking
mechanism. This is ok if only one thread is allowed to push/pop, but
that is not the case.

To fix this, head and tail were given their own mutexes and the fetch
and increment is protected by it.