Weaken memorying ordering throughout
SEQ_CST (total ordering) was used for every atomic operation before the
past few commits. This is the final commit that changes the memory
ordering of all atomic operations.

Most RMW atomic operations were modified to use ACQ_REL memory ordering.
In this case, it essentially does the same thing as SEQ_CST. This memory
ordering is required because other threads can update these values. That
means RELAXED is out of the picture (if we need the most recent value).
ACQUIE or RELEASE could've been used, but we'd like to ensure the
operaton
1. Loads the correct logical value.
2. Stores the correct logical value and allows other ACQUIRE ops to see
   it.

Loads were weakened to ACQUIRE for similar reasons as (1) above, and
stores were weakened to RELEASE for similar reasons as (2).

There were a couple cases where RELAXED made the most sense.
- We didn't care to load the most recent pool_flags because it
  hurts performance and will be laoded with ACQUIRE later.
- A value was never actually loaded, so using RELAXED for store was ok.