Move config from fiber.h to config.mk
I like the idea of having a project's configuration be in a C header
file, but it starts to have problems when you need to conditionally
compile and link to files based on the config. I ran into this issue
when I added options for compiling the fifo queue, gcc and clang atomic
implementation, and pthread implementation.

My solution was to add the source files directly in fiber.c. This
worked fine until I started to write some tests that would be 1000x
easier if I could mock functions (I wanted to mock fiber_libversion
to return a custom value). This is very difficult to do cleanly with my
approach, so I moved all the config/conditional compilation to make.

This centralizes the configuation and allows for more flexibility when
building Fiber. The trade-off is complexity: this adds more garbage to
the Makefile and obscures some config macros.