Change threads_number policy and fix fiber_wait deadlock
Before this commit, the pool could end up with 0 threads by calling
fiber_threads_remove, but the user wasn't allowed to create a pool with
0 threads. I think there are legitimate use cases for initializing a
pool without starting any threads yet, so I decided to unify this
policy.

This decision came after I discovered a bug with fiber_wait. If there
are 0 threads in the pool and the job queue's length is greater than 0,
the caller would wait until the user added more threads. This is
obviously problematic because the caller thread may be the user's only
thread. A seemingly simple fix would have been to require at least one
thread to be in the pool, but, for reasons described above, I don't think
that is the correct choice.

To fix this bug, I removed the case where the queue length being > 0
led to a wait on the sync semaphore. A comment just above it describes
the new condition and why it works.