Remove pool_flags and fix fiber_wait issues
fiber_wait has proved to be a very difficult function because of its
interaction with fiber_threads_remove. Since we atomically update values
like threads_number and threads_working, we can never trust their values
after loading them (they could become stale immediately after the load).
This led to the following possible scenario:
- A user's thread wishes to wait on the jobs in a fiber_pool. It calls
  fiber_wait.
- The user just called fiber_threads_remove (in another thread or before
  calling fiber_wait). This call passed threads_number as the total amount
  of threads in the pool.
- fiber_wait sets the wait flag and fetches threads_number,
  threads_working, and queue_length.
- threads_number was > 0 at the fetch time but is 0 now because the
  previous call to fiber_threads_remove removed all the threads.
  threads_working was also > 0 at the fetch time.
- The threads_sync semaphore is now being waited on, but there are no
  threads to post to it. This is a deadlock.

This change fixes the above sequence by requiring the last thread in a
pool to check and handle the wait flag after decrementing
threads_number. This ensures one of the following is true:
1. fiber_wait_callers > 0 when the thread checks the flag. It will
   handle it.
2. fiber_wait_callers <= 0 (yes it can be < 0 temporarily) when the
   thread checks the flag. In this case, the call to fiber_wait will see
   a threads_num of 0 (assuming no other threads have been added , so it
   will not wait on threads_sync.

This commit also removes the "pool_flags" member of struct fiber_pool.
This member was redundant and unneeded because other values, like
threads_kill_number, stored the state of this flag. A new member was
added to the pool to replace the wait flag: fiber_wait_callers.

Moving the wait flag from a boolean to a number fixed the following
problems:
1. If multiple threads called fiber_wait, it was not guaranteed that all
   of them would eventually wake up. This value allows threads to
   sem_post to threads_sync until all callers to fiber_wait wake up.
2. There were cases where a thread exiting fiber_wait would turn off the
   wait flag just after a thread entering fiber_wait set it. This bug
   would have been rare, but it would result in a deadlock.

This commit also makes queue length functions required. fiber_wait
relies on this function.