fix fiber_worker_wake_other deadlock and redefine ETHREADING errors
fiber_worker_wake_other is a function that pushes a dummy job onto the
job queue to wake up any sleeping threads. This is useful when all
threads are sleeping and the pool must handle a new flag like wait or
kill. This function could result in a deadlock in the following
scenario:
1. The user attempts to kill all threads with fiber_threads_remove but
   specifies a number higher than the current number of threads.
2. The last thread to die will see a value of threads_kill_number > 0 so
   it will call fiber_worker_wake_other.
3. This will push the dummy job onto the queue with the FIBER_QUEUE_BLOCK
   flag.
4. If the queue is full, this call will never return.

The 4th step in the sequence is true because of the following:
If there are no threads to pop jobs off the queue, the queue will remain
full. While the queue remains full, the call to push the dummy job will
never return.

The simple fix for this is to just push the job onto the queue with the
FIBER_QUEUE_NO_BLOCK flag and ensure no error occurred or the error was
due to the queue being full. This error case is ok because if the queue
is full, some thread will eventually pop a job off the queue.

This commit also moved and redefined Fiber's EAGAIN and EINTR errors.
These error were moved to fiber.h so functions like fiber_job_push can
pass these errors from the queue implementation. Prior to this, all
errors from queue_ops.push were mapped to FBR_EPUSH_JOB.

These errors are no longer defined based on errno's EAGAIN and EINTR.
This was just stupid.