Remove queue and threading mocks, add alloc_trace mock
The queue and threading noop mocks were dumb. They are both very simple
modules so mocking them shouldn't be necessary.

I added a new "mock" called alloc_trace and used it in tests for
fiber_init. Mock may be a misnomer because it is more of an extension to
malloc/free for testing.

alloc_trace tracks each call to malloc and free to ensure each pointer
returned by malloc is freed. It also ensures the number of times each
function is called is equal. alloc_trace uses a global state that is
reset before test case, and the verify function should be called
directly after the test case. Fiber can call malloc and free from
multiple threads in parallel, so alloc_trace uses a mutex to protect
its global state in those functions.

alloc_trace's implementation is super bad, but it works. I may update it
in the future if it slows tests down or I need to track more pointers.