Remove setting thread count in fiber_init to match 1e9065c
In the commit listed above, I moved the threads_number addition from
fiber_threads_add to the actual thread worker function. I forgot to
remove the part in fiber_init that sets the threads_number.

This change didn't cause fiber_init tests to fail because the threads
didn't make it to the part that incremented the count before we checked
the value. This also caused the test to fail after fixing the
threads_number store in fiber_init.

My solution is to busy wait until threads_number matches the
initialization value or 1.5 seconds have passed (usleep and nanosleep
aren't available, and I'm not waiting a whole second).

I will not add code in fiber_init or fiber_threads_add to wait until the
threads have started to return. This is not a race condition unless you
are creating pools and destroying them very, very quickly. That is not
how Fiber should be used so I'm ok with that.