Embed queue_ops in pool
Seems silly to malloc and free queue_ops separately from the pool when
its lifetime is the same. The only argument I could see in favor of
malloc is if it was a large and/or infrequently accessed member that
polluted the cache when accessing pool. queue_ops is used very
frequently and is only the size of 5 pointers.