Improve fiber_workers_cancel performance
I knew calling fiber_thread_cancel and fiber_thread_join back to back
would result in terrible performance, but I wasn't expecting it to be
absolutely garbage.

In a test with 32 threads I found the function took ~13ms. This may not
seem that bad, but that was about 13% of the program's runtime
(src/example.c with 32 threads only). I knew I could do better with very
little effort so I did. The new version takes ~4ms, which is only 4.6%
of the same program's runtime. Low hanging fruit.

In order to do this an array of tids had to be allocated. I kept the old
implementation as a fallback if malloc returns NULL.