Add new fiber_job_push_raw function
A function __fiber_job_push was used internally to push a job onto the
queue without assigning it an ID. I decided to add it to the public API
because I can see a use case where the user may want to assign their own
job IDs.

I also reverted the version to 0.2.0. This is probably bad but nobody is
using this project so it shouldn't matter. I want to tag a new release
but thought it would be weird to skip many minor versions.