Add fiber_atomic and delete current wait implementation
I needed atomic operations for another project so I wrote fiber_atomic.
No point in keeping two sets of portable atomic operations source code
so I added the other as a submodule. This new project uses macros
instead of functions so any type can use them.

I also noticed many problems with the current fiber_wait implementation
so I removed the code for it. The code works great if the sem_wait path
is taken by the waiter. When that path is not taken, it is possible for
the pool to post to the sem since we incremented fiber_wait_callers
before fetching the thread/queue numbers. It is also possible that no
thread posts the sem. In this case, fiber_wait_callers would not be
decrement and wouldn't reflect the real number of waiters. It is hard
for the waiter to know which happened.

I could add code to reconcile the two numbers if the sem_wait path is
not taken but this leads to another problem. The semaphore's value is
one more than it is supposed to be while the waiter is fixing its value.
It is possible that another thread calls fiber_wait, takes the sem_wait
path (which means there is work to be done in the pool), and instantly
returns because the semaphore's value is incorrect. This is obviously
unacceptable.