Add bitstring implementation & begin fiber_wait
I've decided to implement a lock free singly linked list that uses EBR
to reclaim nodes. This will be used to allow fiber_wait callers to add
their "intent" to wakeup after all the jobs in the pool are completed.
This may or may not be the most efficient way to implement this "intent"
list but I want to try it.

To implement this list, each thread requires some local storage to put
nodes that are logically deleted but not yet freed. To implement this
part I am going to create a table and give each thread an entry in that
table to hold TL storage. I am going to use a bitstring to track which
entries in the table are freed and which are used.