Add aligned_alloc & aligned_free functions
Aligning structures can be very useful in many scenarios. Sometimes it's
required, preferred, or helps peformance. In my case, I'd like to align
some thread local structures on cache line boundaries to prevent false
sharing.

aligned_malloc is built on top of a general purpose allocator. It simply
overallocates and aligns the pointer returned to the user. This
implementation is not the best, but it is portable and works with any
general purpose allocator.