Fix and change free logic
When fbr_free() is called we no longer use fbr_thread_cancel() to tell
the threads to cancel. Instead we mark the pool inactive, set the
thread_kill_number to a high number, and wake up threads that are
possibly sleeping on the queue futex. If there are no threads in the
pool we still synchronously clean everything up.