Use user provided buffer on init if provided
One of Fiber's selling points is no memory allocation after
initialization (except for pthread_create's allocations). To make this
point even more prevalent I altered the way memory is allocated for the
pool.

This commit allows the user to either:
1. Give fbr_init a buffer of buffer_size. If it is big enough for the
   pool, Fiber will use it.
2. Don't give fbr_init a buffer but give it an allocator to allocate the
   buffer.

To facilitate this change I added a function to get the minimum size a
pool requires given a set of initialization options. I wish I could
somehow do this at compile time so the user can use a static buffer, but
it's not very feasible currently.

One note is that the job queue implementation may still require an allocator
if it is a dynamic queue. A queue implementation is allowed to specify
the size it requires, but the allocator is still passed to it. There can
be cases where fbr_init does not require an allocator (because the user
provided a buffer), but the queue implementation does.