Fix: fbr_free() race when called after spawning threads
If a user called fbr_init() or fbr_thread_add() and fbr_free() directly
after, it is possible that the pool could be completely freed before the
thread even entered its main function.

fbr_free() only checked pool->thread_num when determining if it could
free the pool synchronously. It is possible that a thread was created
but the thread num remained 0 due to it not being scheduled. This leads
to a race where the pool can be freed out from under a thread while it
is still initializing.

To fix this, I added a new counter variable named "thread_spawning_num"
that tracks how many threads are currently being initialized. fbr_free()
will not free the pool until it sees that this value is 0.